Moved compute to interface
