+++ include/compute/interface.hpp
@@ -1,0 +1,10 @@
+#ifndef _COMPUTE_INTERFACE_HPP
+#define _COMPUTE_INTERFACE_HPP
+
+#include "compute.hpp"
+#include "compute_partial_derivatives.hpp"
+#include "compute_properties.hpp"
+#include "compute_receivers.hpp"
+#include "compute_sources.hpp"
+
+#endif
