+++ src/compute/compute.cpp
@@ -1,9 +1,9 @@
-#include "../include/compute.h"
-#include "../include/jacobian.h"
-#include "../include/kokkos_abstractions.h"
-#include "../include/quadrature.h"
-#include "../include/shape_functions.h"
-#include "../include/specfem_setup.hpp"
+#include "compute/interface.hpp"
+#include "jacobian.h"
+#include "kokkos_abstractions.h"
+#include "quadrature.h"
+#include "shape_functions.h"
+#include "specfem_setup.hpp"
 #include <Kokkos_Core.hpp>
 #include <tuple>
 #include <vector>
