+++ src/compute/compute_properties.cpp
@@ -1,8 +1,8 @@
-#include "../include/compute.h"
-#include "../include/jacobian.h"
-#include "../include/kokkos_abstractions.h"
-#include "../include/shape_functions.h"
-#include "../include/specfem_setup.hpp"
+#include "compute/interface.hpp"
+#include "jacobian.h"
+#include "kokkos_abstractions.h"
+#include "shape_functions.h"
+#include "specfem_setup.hpp"
 #include <Kokkos_Core.hpp>
 
 specfem::compute::properties::properties(const int nspec, const int ngllz,
