+++ tests/unittests/source/source_location_tests.cpp
@@ -1,13 +1,13 @@
-#include "../../../include/compute.h"
-#include "../../../include/material.h"
-#include "../../../include/mesh.h"
-#include "../../../include/params.h"
-#include "../../../include/read_mesh_database.h"
-#include "../../../include/read_sources.h"
-#include "../../../include/source.h"
-#include "../../../include/specfem_setup.hpp"
 #include "../Kokkos_Environment.hpp"
 #include "../MPI_environment.hpp"
+#include "compute/interface.hpp"
+#include "material.h"
+#include "mesh.h"
+#include "params.h"
+#include "read_mesh_database.h"
+#include "read_sources.h"
+#include "source.h"
+#include "specfem_setup.hpp"
 #include "yaml-cpp/yaml.h"
 #include <stdexcept>
 #include <string>
